internet status view to show connected or disconnected status
