+++ Src/Controls/Styles/InternetStatusStyle.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+export const InternetContainerViewStyles = StyleSheet.create({
+  networkViewContainer: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    height: 60,
+    width: '100%',
+    backgroundColor: '#F02525',
+  },
+});
