+++ Src/Controls/Styles/index.ts
@@ -4,3 +4,4 @@ export { NeuroAccessBackgroundStyle } from "./NeuroAccessBackgroundStyle";
 export { ShowLabelsForAuthStyle } from "./ShowLabelsForAuthStyle";
 export { SplashBackgroundStyle } from "./SplashBackgroundStyle";
 export { TextLabelStyle } from "./TextLabelStyle";
+export {InternetContainerViewStyles} from './InternetStatusStyle';
