+++ Src/Controls/index.ts
@@ -1,10 +1,11 @@
-export { ActionButton } from "./ActionButton";
-export { ChooseNeuroAccessAppContext } from "./ChooseNeuroAccessAppContext";
-export {default as InputBox} from "./InputBox";
-export {Loader} from './Loader';
+export { ActionButton } from './ActionButton';
+export { ChooseNeuroAccessAppContext } from './ChooseNeuroAccessAppContext';
+export { default as InputBox } from './InputBox';
+export { Loader } from './Loader';
 export { NavigationHeader } from './NavigationHeader';
-export { NeuroAccessBackground } from "./NeuroAccessBackground";
-export { NeuroStatusBar } from "./NeuroStatusBar";
-export { ShowLabelsForAuth } from "./ShowLabelsForAuth";
-export { SplashBackground } from "./SplashBackground";
-export { TextLabel } from "./TextLabel";
+export { NeuroAccessBackground } from './NeuroAccessBackground';
+export { NeuroStatusBar } from './NeuroStatusBar';
+export { ShowLabelsForAuth } from './ShowLabelsForAuth';
+export { SplashBackground } from './SplashBackground';
+export { TextLabel } from './TextLabel';
+export { InternetStatus } from '@Controls/InternetStatus';
