+++ Src/Translations/Resources/en.ts
@@ -24,5 +24,9 @@ export default {
   settingScreen:{
     selectLanguage:' Select Language ',
     changeTheme:'Change Theme'
+  },
+  internetSatus: {
+    connected: 'Internet Connected',
+    disconnected: 'Internet Not Connected!'
   }
 };
